modularised slash coms, not working
