+++ moderation.js
@@ -34,5 +34,4 @@ async function moderateMessage(openai, message, naughtyWords, banList) {
 }
 
 // Export the moderation function
-module.exports = moderateMessage;
 module.exports = { moderateMessage };
